usage example with scripts, inputs and outputs
This example contains a script that runs the inference for a single
Drosophila chromosome.
Inputs: annotations of conserved regions and putatively functional
substitutions, genetic maps, neutral polymorphism, proxy for spatial
variation in mutation rates and a mask of codons used by the inference.
Outputs: pre-inference calculated grid elements (i.e. maps of B and S
for each fitness effect on the grids of DFEs), a text file with inferred
parameters and statistics and a map of the effects of linked selection.